Moved code for generating scaladoc to project/Doc.scala.
